config.sh safeguard - no screenrc install if no screen detected

+++ config.sh
@@ -1,18 +1,25 @@
 #!/bin/bash
 
 # nano, screen and bash
-echo "linking nano, screen, and bash configuration..."
+echo "linking nano, bash, and xresources configuration..."
 ln -f $(pwd)/.nanorc ~
-ln -f $(pwd)/.screenrc ~
-ln -f $(pwd)/.bash_profile ~
 ln -f $(pwd)/.bashrc ~
 ln -f $(pwd)/.xresources ~
-
+# screen
+if type screen 2>/dev/null; then
+  echo "screen detected, adding configuration..."
+  ln -f $(pwd)/.screenrc ~
+  ln -f $(pwd)/.bash_profile ~
+fi
 
 # terminal font
-echo "copying console-setup..."
-sudo rm /etc/default/console-setup
-sudo cp console-setup /etc/default/console-setup
+echo "setup console-setup? (requires root) [y/n]"
+read -s -n 1 -r char
+if [ $char = "y" ]; then
+  echo "copying console-setup..."
+  sudo rm /etc/default/console-setup
+  sudo cp console-setup /etc/default/console-setup
+fi
 
 #i3
 echo "setup i3? [y/n]"
